Rename DiscoverSessionView to SingleTabView
