+++ src/platform/plugins/shared/discover/public/application/main/components/single_tab_view/branded_loading_indicator.tsx
