+++ src/platform/plugins/shared/discover/public/application/main/components/single_tab_view/index.ts
@@ -9,4 +9,4 @@
 
 export { BrandedLoadingIndicator } from './branded_loading_indicator';
 export { NoDataPage } from './no_data_page';
-export { DiscoverSessionView, type DiscoverSessionViewProps } from './session_view';
+export { SingleTabView, type SingleTabViewProps } from './single_tab_view';
