+++ src/platform/plugins/shared/discover/public/application/main/components/single_tab_view/main_app.tsx
