+++ src/platform/plugins/shared/discover/public/application/main/components/single_tab_view/no_data_page.tsx
