+++ src/platform/plugins/shared/discover/public/application/main/components/single_tab_view/redirect_not_found.tsx
