+++ src/platform/plugins/shared/discover/public/application/main/components/single_tab_view/single_tab_view.tsx
@@ -51,7 +51,7 @@ import { DiscoverMainApp } from './main_app';
 import { useAsyncFunction } from '../../hooks/use_async_function';
 import { ScopedServicesProvider } from '../../../../components/scoped_services_provider';
 
-export interface DiscoverSessionViewProps {
+export interface SingleTabViewProps {
   customizationContext: DiscoverCustomizationContext;
   customizationCallbacks: CustomizationCallback[];
   urlStateStorage: IKbnUrlStateStorage;
@@ -69,13 +69,13 @@ type InitializeSession = (options?: {
   shouldClearAllTabs?: boolean;
 }) => Promise<SessionInitializationState>;
 
-export const DiscoverSessionView = ({
+export const SingleTabView = ({
   customizationContext,
   customizationCallbacks,
   urlStateStorage,
   internalState,
   runtimeStateManager,
-}: DiscoverSessionViewProps) => {
+}: SingleTabViewProps) => {
   const dispatch = useInternalStateDispatch();
   const services = useDiscoverServices();
   const { core, history, getScopedHistory } = services;
