+++ src/platform/plugins/shared/discover/public/application/main/components/tabs_view/tabs_view.tsx
@@ -9,7 +9,7 @@
 
 import { UnifiedTabs, type UnifiedTabsProps } from '@kbn/unified-tabs';
 import React, { useCallback } from 'react';
-import { DiscoverSessionView, type DiscoverSessionViewProps } from '../session_view';
+import { SingleTabView, type SingleTabViewProps } from '../single_tab_view';
 import {
   createTabItem,
   internalStateActions,
@@ -21,7 +21,7 @@ import {
 import { useDiscoverServices } from '../../../../hooks/use_discover_services';
 import { usePreviewData } from './use_preview_data';
 
-export const TabsView = (props: DiscoverSessionViewProps) => {
+export const TabsView = (props: SingleTabViewProps) => {
   const services = useDiscoverServices();
   const dispatch = useInternalStateDispatch();
   const items = useInternalStateSelector(selectAllTabs);
@@ -40,7 +40,7 @@ export const TabsView = (props: DiscoverSessionViewProps) => {
   );
 
   const renderContent: UnifiedTabsProps['renderContent'] = useCallback(
-    () => <DiscoverSessionView key={currentTabId} {...props} />,
+    () => <SingleTabView key={currentTabId} {...props} />,
     [currentTabId, props]
   );
