+++ src/platform/plugins/shared/discover/public/constants.ts
@@ -12,5 +12,4 @@ export const ADHOC_DATA_VIEW_RENDER_EVENT = 'ad_hoc_data_view';
 export const SEARCH_SESSION_ID_QUERY_PARAM = 'searchSessionId';
 
 // TEMPORARY: This is a temporary flag to enable/disable tabs in Discover until the feature is fully implemented.
-export const TABS_ENABLED =
-  window?.localStorage?.getItem('discoverExperimental:tabs') === 'true' || false;
+export const TABS_ENABLED = true;
